Move misplaced command file
