+++ editor/src/components/canvas/canvas-strategies/reparent-utils.ts
@@ -26,7 +26,7 @@ import { forceNotNull } from '../../../core/shared/optional-utils'
 import { addImportsToFile } from '../commands/add-imports-to-file-command'
 import { BuiltInDependencies } from '../../../core/es-modules/package-manager/built-in-dependencies-list'
 import { CSSCursor } from '../canvas-types'
-import { addToReparentedToPaths } from './add-to-reparented-to-paths-command'
+import { addToReparentedToPaths } from '../commands/add-to-reparented-to-paths-command'
 import { getStoryboardElementPath } from '../../../core/model/scene-utils'
 import { getUtopiaID } from '../../../core/model/element-template-utils'
 import { addElement } from '../commands/add-element-command'
